Move CostsEntryScreen to a separate directory
